+++ smart-home-planner/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.4.0 - 
 
 ### Added
-- New Excluded Devices section on Settings
+- New "Excluded Devices" section on Settings
 - More automatic devices exclusions when syncing with HA
 - Support to open a device in HA
 - Now editing the name of a device in the app also edits it in HA